+++ changelog.md
@@ -7,3 +7,4 @@
 - fixed compatibility crash with MineColonies (reported by @Ryus Lightworks)
 - fixed bug with Onyx not granting advancement
 - fixed bug with current loot bonus displaying invalid value
+- fixed compatibility issue with accessories having range bonus after using crafting table